+++ server/themes/default/public/javascripts/admin/admin.main.js
@@ -24,6 +24,9 @@ angular.module('app', ['ngRoute', 'ngResource', 'ngSanitize', 'angular-uuid', 'u
         AliasExport: $resource('/api/capcodeExport', null, {
           'post': { method:'POST', isArray: false }
         }),
+        AliasImport: $resource('/api/capcodeImport', null, {
+          'post': { method:'POST', isArray: false }
+        }),
       };
     }])
     
@@ -105,6 +108,87 @@ angular.module('app', ['ngRoute', 'ngResource', 'ngSanitize', 'angular-uuid', 'u
         });
       };
 
+      $scope.aliasImport = function () {
+        var modalHtml = '<div class="modal-header"><h5 class="modal-title" id="modal-title">Impot Aliases</h5></div>';
+        var messages = `<p>Available Columns: address, alias, agency, color, icon, ignore, pluginconf</p>
+                        <p>Required columns are "address" and "alias", all others are optional.</p>`;
+        modalHtml += '<div class="modal-body"><p><input type="file" id="importcsv"/></p><p>CSV file to be imported</p>' + messages + '</div>';
+        modalHtml += '<div class="modal-footer"><button class="btn btn-success" ng-click="confirmImport()">Import</button><button class="btn btn-danger" ng-click="cancelImport()">Cancel</button></div>';
+        var modalInstance = $uibModal.open({
+          template: modalHtml,
+          controller: ImportController,
+
+        });
+        modalInstance.result.then(function () {
+          $scope.aliasImportConfirmed();
+        }, function () {
+          //$log.info('Modal dismissed at: ' + new Date());
+        });
+      };
+
+      $scope.aliasImportConfirmed = function () {
+        $scope.loading = true;
+        var filename = document.getElementById("importcsv");
+        if (filename.value.length < 1) {
+          // noidea i stole this code. 
+        } else {
+          var file = filename.files[0];
+          console.log(file)
+          var fileSize = 0;
+          if (filename.files[0]) {
+            var reader = new FileReader();
+            reader.onload = function (e) {
+              var rows = e.target.result.split("\n");
+              Api.AliasImport.post(rows).$promise.then(function (response) {
+                console.log(response)
+                $scope.loading = false;
+                $scope.results = response.results
+                var resultModalHtml = '<div class="modal-header"><h5 class="modal-title" id="modal-title">Import Results</h5></div>';
+                resultModalHtml += `<div class="modal-body">  
+                    <table class="table table-striped">
+                       <thead>
+                       <tr>
+                          <th>Address</th>
+                          <th>Alias</th>
+                          <th>Result</th>
+                        </tr>
+                        </thead>
+                        <tbody>
+                        <tr ng-repeat="result in results">
+                          <td>{{ result.address }}</td>
+                          <td>{{ result.alias }}</td>
+                          <td>{{ result.result }}</td>
+                        </tr>
+                        </tbody>
+                      </table>
+                      </div>`
+                resultModalHtml += '<div class="modal-footer"><button class="btn btn-success" ng-click="okImport()">OK</button></div>';
+                var modalInstance = $uibModal.open({
+                  template: resultModalHtml,
+                  controller: ImportController,
+                  scope: $scope
+                });
+
+              }, function (response) {
+                $scope.loading = false;
+                console.log(response)
+                var resultModalHtml = '<div class="modal-header"><h5 class="modal-title" id="modal-title">Import Failed</h5></div>';
+                resultModalHtml += `<div class="modal-body">  
+                    <p>Failed to Parse CSV file, please check the file and try again!</p>
+                    `
+                resultModalHtml += '<div class="modal-footer"><button class="btn btn-success" ng-click="okfailedImport()">OK</button></div>';
+                var modalInstance = $uibModal.open({
+                  template: resultModalHtml,
+                  controller: ImportController,
+                  scope: $scope
+                });
+              })
+            }
+            reader.readAsText(filename.files[0]);
+          }
+          return false; //no idea what this does, was also in the code i stole and it doesn't work without it. 
+        }
+      };
 
       $scope.messageDetail = function(address) {
           $location.url('/aliases/'+address);
@@ -175,14 +259,35 @@ angular.module('app', ['ngRoute', 'ngResource', 'ngSanitize', 'angular-uuid', 'u
         });
       };
       
-      var ConfirmController = function($scope, $uibModalInstance) {
-        $scope.confirmDelete = function() {
+      var ConfirmController = function ($scope, $uibModalInstance) {
+        $scope.confirmDelete = function () {
           $uibModalInstance.close();
         };
-      
-        $scope.cancelDelete = function() {
+
+        $scope.cancelDelete = function () {
           $uibModalInstance.dismiss('cancel');
         };
+      };
+
+      var ImportController = function ($scope, $uibModalInstance) {
+        $scope.confirmImport = function () {
+          $uibModalInstance.close();
+        };
+        $scope.cancelImport = function () {
+          $uibModalInstance.dismiss('cancel');
+        };
+        $scope.okImport = function () {
+          $uibModalInstance.close();
+          $scope.aliasRefreshRequired = 1
+          $scope.alertMessage.text = 'Alias refresh required!';
+          $scope.alertMessage.type = 'alert-warning';
+          $scope.alertMessage.show = true;
+          $timeout(function () { $scope.alertMessage.show = false; }, 3000);
+          $location.url('/aliases/');
+        }
+        $scope.okfailedImport = function () {
+          $uibModalInstance.close();
+        }
       };
     }])